split qwen36 moe dry run reporting
