+++ crates/runner/src/main.rs
@@ -14,6 +14,7 @@ mod prefill_engine;
 mod qwen35_dflash_engine;
 mod qwen35_runtime;
 mod qwen36_moe_decode;
+mod qwen36_moe_dry_run;
 mod qwen36_moe_engine;
 mod qwen36_moe_host;
 mod qwen36_moe_layers;
