+++ crates/runner/src/qwen36_moe_engine.rs
@@ -10,29 +10,24 @@
 //! the only meaningful runtime check is "did the safetensors index match
 //! what we expect from the config" plus a budget comparison.
 
-use std::collections::BTreeMap;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 use anyhow::{anyhow, Context, Result};
-use gpu_hal::{set_backend, Backend, GpuBuffer, ScalarType, VirtualAllocationRole, VirtualArena};
-use model_store::manifest::LayoutTag;
+use gpu_hal::{set_backend, Backend, GpuBuffer, ScalarType, VirtualArena};
 use model_store::BakedStore;
-use qwen36_moe::config::{Config, TextConfig};
-use qwen36_moe::loader::{ScalarKind, WeightLoader};
-use qwen36_moe::state::{StateAccount, StateLayout};
-use qwen36_moe::weights::{
-    checkpoint_dtype_acceptable, expected_tensor_specs, CheckpointAccount, CheckpointDtype,
-};
+use qwen36_moe::config::TextConfig;
 
 use crate::qwen36_moe_decode::{
     argmax_bf16_logits, host_final_norm_lm_head, run_chained_decode, run_chained_decode_fast,
     run_chained_decode_fast_with_expert_prefetch, sample_bf16_logits, ExpertPrefetchPhase,
     ExpertRoute, LayerBuffers, MtpLayerBuffers, MultiLayerGeom, XorshiftRng,
 };
+use crate::qwen36_moe_dry_run::{
+    print_report, run_qwen36_moe_dry_run, ContextSizeSource, DryRunReport,
+};
 use crate::qwen36_moe_host::{host_load_bytes, load_lm_head_bf16, lookup_embed_row};
 use crate::qwen36_moe_layers::{
-    load_all_layer_buffers, load_layer_buffers, load_to_gpu, resolve_qwen36_store_name,
-    store_contains_qwen36, store_layout_qwen36, Qwen36WeightMode,
+    load_all_layer_buffers, load_layer_buffers, load_to_gpu, Qwen36WeightMode,
 };
 use crate::qwen36_moe_prefetch::handle_moe_expert_prefetch;
 use crate::qwen36_moe_residency::{MoeExpertResidencyConfig, MoeExpertResidencyManager};
@@ -49,726 +44,10 @@ use crate::qwen36_moe_vmm::{
     moe_island_prefetch_transition_min_observations, should_try_moe_expert_vmm,
     should_use_qwen36_kv_vmm, virtual_kv_stats_for_layers, MoeExpertVmmMode,
 };
-use crate::registry::{FamilyParams, Qwen36MoeKernelParams, RegistryEntry};
+use crate::registry::{Qwen36MoeKernelParams, RegistryEntry};
 
-const GIB: f64 = (1024 * 1024 * 1024) as f64;
 const MIB: f64 = (1024 * 1024) as f64;
 const QWEN36_NUM_SPECULATIVE_TOKENS: usize = 3;
-
-pub struct DryRunReport {
-    pub config: Config,
-    pub kernel_params: Qwen36MoeKernelParams,
-    pub checkpoint: CheckpointAccount,
-    pub int4_projected_bytes: u64,
-    pub state: StateAccount,
-    pub on_disk_bytes: Option<u64>,
-    pub on_disk_tensor_count: Option<usize>,
-    pub missing_tensors: Vec<String>,
-    pub dtype_mismatches: Vec<String>,
-    /// Set when safetensors files are partially present but the loader
-    /// couldn't open them (e.g. mid-download). The analytic accounting is
-    /// still emitted so the user gets useful output.
-    pub loader_warning: Option<String>,
-    pub registry_budget_bytes: u64,
-    pub gpu_total_vram_bytes: u64,
-    /// Populated when an INT4 baked package is present and loadable.
-    pub bake: Option<BakeAccount>,
-    /// What the dry-run actually used for KV-cache sizing — and where it
-    /// came from. The string flavor (`Explicit`, `EstimatedFromPrompt`,
-    /// `MaxNewTokensOnly`) lets `print_report` flag the worst case so the
-    /// user knows the `fit:YES` answer covered their realistic prompt.
-    pub context_size_used: usize,
-    pub context_size_source: ContextSizeSource,
-}
-
-/// How `context_size_used` was derived. Anything other than `Explicit`
-/// means the dry-run is making an assumption the caller should verify.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ContextSizeSource {
-    /// User passed `--context-size`; honoured verbatim.
-    Explicit,
-    /// User passed `--prompt` but no `--context-size`; estimated as
-    /// (prompt char count) + max_new_tokens. Char count is a conservative
-    /// upper bound on token count for English-ish text — true tokenisation
-    /// would need the model's tokenizer, which the dry-run path doesn't
-    /// load.
-    EstimatedFromPrompt,
-    /// Neither `--context-size` nor `--prompt` was given. Defaults to
-    /// max_new_tokens — almost certainly an undercount for any real
-    /// session. The report flags this as a worst-case caveat.
-    MaxNewTokensOnly,
-}
-
-/// Summary of a baked-package's contents — the runtime-ready view the
-/// kernel will see at decode time. Built by mmap'ing the bake at
-/// `model-store::bake_dir_int4()` and walking its manifest.
-#[derive(Debug, Clone)]
-pub struct BakeAccount {
-    pub bake_dir: PathBuf,
-    pub manifest_format_version: u32,
-    pub manifest_converter_version: u32,
-    pub tensor_count: usize,
-    pub weights_bin_bytes: u64,
-    /// Per-LayoutTag tensor counts. Sorted alphabetically by tag name for
-    /// deterministic output.
-    pub by_layout: BTreeMap<String, usize>,
-    /// Tensors expected per `expected_tensor_specs` that are absent from
-    /// the bake's index. A non-empty list means the bake is incomplete.
-    pub missing_specs: Vec<String>,
-    /// INT4 expert tensors per layer the runtime relies on. Each entry is
-    /// `(layer_idx, name)` for any of `gate_up_proj`, `down_proj` that's
-    /// missing or has a wrong layout.
-    pub bad_expert_tensors: Vec<(usize, String)>,
-    /// Cached aggregate byte sizes per category. Useful for comparing the
-    /// bake against the analytic INT4 projection.
-    pub int4_quantized_bytes: u64,
-    pub raw_bytes: u64,
-    /// Optional live VMM probe populated when requested by environment.
-    /// This loads real baked tensors into role-tagged virtual allocations
-    /// without switching decode descriptors over to VMM yet.
-    pub vmm_probe: Option<VirtualBakeProbe>,
-}
-
-#[derive(Debug, Clone)]
-pub struct VirtualBakeProbe {
-    pub allocations: usize,
-    pub logical_bytes: usize,
-    pub reserved_bytes: usize,
-    pub resident_bytes: usize,
-    pub logical_resident_bytes: usize,
-    pub mapping_count: usize,
-    pub loaded_tensors: Vec<String>,
-    pub failed_tensors: Vec<String>,
-}
-
-pub fn run_qwen36_moe_dry_run(
-    model_dir: &Path,
-    entry: &RegistryEntry,
-    total_vram: u64,
-    context_size: usize,
-    context_size_source: ContextSizeSource,
-    batch_size: usize,
-    kv_fp8: bool,
-    no_bake: bool,
-    ordinal: usize,
-) -> Result<DryRunReport> {
-    let kernel_params = match entry.params {
-        FamilyParams::Qwen36Moe(p) => p,
-        _ => return Err(anyhow!("registry entry is not Qwen36Moe family")),
-    };
-
-    let config = qwen36_moe::config::load_config(model_dir)
-        .map_err(|e| anyhow!("parse config.json: {e}"))?;
-
-    sanity_check_kernel_params(&config.text_config, &kernel_params)?;
-
-    let weight_prefix = kernel_params.weight_prefix;
-    let specs = expected_tensor_specs(&config.text_config, weight_prefix);
-    let checkpoint = CheckpointAccount::from_config(&config.text_config);
-    let int4_projected_bytes = checkpoint.project_int4_total_bytes(&config.text_config, 128);
-
-    // The sidecar window is consulted only when --kv-fp8 is on AND the
-    // env-var helper is enabled. The persistent kernel treats it as a
-    // rolling recent-token window.
-    let sidecar_window = if kv_fp8 && qwen35::state::kv_fp8_bf16_sidecar_enabled() {
-        let window = qwen35::state::kv_fp8_bf16_sidecar_window_tokens()
-            .unwrap_or(context_size)
-            .min(context_size);
-        (window > 0).then_some(window)
-    } else {
-        None
-    };
-    let layout = StateLayout::new(context_size, batch_size, kv_fp8, sidecar_window);
-    let state = StateAccount::from_config(&config.text_config, layout);
-
-    let bake = inspect_bake(model_dir, &config.text_config, weight_prefix, ordinal);
-
-    let mut on_disk_bytes = None;
-    let mut on_disk_tensor_count = None;
-    let mut missing_tensors = Vec::new();
-    let mut dtype_mismatches = Vec::new();
-    let mut loader_warning = None;
-    if !no_bake_only_safetensors(model_dir, no_bake) {
-        // No safetensors expected — bake-only path. PR 3 doesn't yet open
-        // the baked store; revisit once PR 6 lands the INT4 reader.
-    } else if !has_safetensors(model_dir) {
-        // Safetensors missing AND no bake support yet — caller probably
-        // needs to download. Surface a helpful note rather than failing,
-        // because the analytic numbers are still useful.
-        loader_warning = Some(format!(
-            "no safetensors in {} — analytic accounting only",
-            model_dir.display()
-        ));
-    } else {
-        match WeightLoader::from_dir(model_dir) {
-            Ok(loader) => {
-                on_disk_tensor_count = Some(loader.tensor_count());
-                let mut total: u64 = 0;
-                let mut io_failures: usize = 0;
-                for spec in &specs {
-                    match loader.meta(&spec.name) {
-                        Ok(meta) => {
-                            total += meta.byte_size();
-                            let got_dtype = scalar_to_checkpoint_dtype(meta.dtype);
-                            if let Some(got) = got_dtype {
-                                if !checkpoint_dtype_acceptable(spec.role, got) {
-                                    dtype_mismatches.push(format!(
-                                        "{} got {:?} (not acceptable for {:?})",
-                                        spec.name, got, spec.role
-                                    ));
-                                }
-                            } else {
-                                dtype_mismatches.push(format!(
-                                    "{} got non-float dtype {:?}",
-                                    spec.name, meta.dtype
-                                ));
-                            }
-                        }
-                        Err(qwen36_moe::loader::LoadError::NotFound(_)) => {
-                            missing_tensors.push(spec.name.clone());
-                        }
-                        // Tensor is in the index but reading metadata failed
-                        // (e.g. shard file not yet present mid-download).
-                        // Don't kill the whole dry-run.
-                        Err(_) => io_failures += 1,
-                    }
-                }
-                on_disk_bytes = Some(total);
-                if io_failures > 0 {
-                    loader_warning = Some(format!(
-                        "{io_failures} tensors could not be read (likely partial download); \
-                         on-disk total is incomplete"
-                    ));
-                }
-            }
-            Err(e) => {
-                loader_warning = Some(format!(
-                    "could not open safetensors from {}: {e}",
-                    model_dir.display()
-                ));
-            }
-        }
-    }
-
-    let kv_bytes_per_token = config
-        .text_config
-        .kv_bytes_per_token(if kv_fp8 { 1 } else { 2 });
-    let registry_budget_bytes = entry.vram.estimate_total(context_size, kv_bytes_per_token);
-
-    Ok(DryRunReport {
-        config,
-        kernel_params,
-        checkpoint,
-        int4_projected_bytes,
-        state,
-        on_disk_bytes,
-        on_disk_tensor_count,
-        missing_tensors,
-        dtype_mismatches,
-        loader_warning,
-        registry_budget_bytes,
-        gpu_total_vram_bytes: total_vram,
-        bake,
-        context_size_used: context_size,
-        context_size_source,
-    })
-}
-
-/// Open the INT4 baked package (if present and valid) and summarise its
-/// contents. Returns None when no bake is on disk or its manifest can't be
-/// parsed; the dry-run continues either way.
-fn inspect_bake(
-    model_dir: &Path,
-    text_config: &TextConfig,
-    weight_prefix: &str,
-    ordinal: usize,
-) -> Option<BakeAccount> {
-    let bake_dir = model_store::bake_dir_int4(model_dir);
-    if !bake_dir.exists() {
-        return None;
-    }
-    // Re-parse the manifest header even though BakedStore does too — we
-    // expose format/converter versions in the report so a stale bake is
-    // visible without having to dig into the file by hand.
-    let manifest_text = std::fs::read_to_string(model_store::manifest_path(&bake_dir)).ok()?;
-    let manifest: model_store::manifest::Manifest = serde_json::from_str(&manifest_text).ok()?;
-    let store = BakedStore::open(&bake_dir).ok()?;
-    let weights_bin_bytes = std::fs::metadata(model_store::weights_bin_path(&bake_dir))
-        .ok()?
-        .len();
-
-    // Aggregate per-layout tensor counts and per-category byte totals.
-    let mut by_layout: BTreeMap<String, usize> = BTreeMap::new();
-    let mut int4_quantized_bytes: u64 = 0;
-    let mut raw_bytes: u64 = 0;
-    for t in &manifest.tensors {
-        let key = format!("{:?}", t.layout);
-        *by_layout.entry(key).or_default() += 1;
-        match t.layout {
-            LayoutTag::Int4Quantized => int4_quantized_bytes += t.byte_len,
-            LayoutTag::Raw => raw_bytes += t.byte_len,
-            _ => {}
-        }
-    }
-
-    // Cross-check the bake against the runtime's expected tensor specs.
-    // A missing entry here would crash the loader at decode time; surface
-    // it during dry-run instead.
-    let specs = expected_tensor_specs(text_config, weight_prefix);
-    let mut missing_specs = Vec::new();
-    for spec in &specs {
-        if !store_contains_qwen36(&store, &spec.name) {
-            missing_specs.push(spec.name.clone());
-        }
-    }
-
-    // The fused MoE expert tensors are the bulk of the bake (~80% of the
-    // INT4 byte total on 35B-A3B). Spot-check every layer; absence here is
-    // fatal because the kernel can't run without a routed expert.
-    let mut bad_expert_tensors = Vec::new();
-    for li in 0..text_config.num_hidden_layers as usize {
-        for kind in ["gate_up_proj", "down_proj"] {
-            let name = format!("{weight_prefix}.layers.{li}.mlp.experts.{kind}");
-            match store_layout_qwen36(&store, &name) {
-                Some(LayoutTag::Int4Quantized) => {}
-                Some(other) => bad_expert_tensors.push((li, format!("{name} (layout={other:?})"))),
-                None => bad_expert_tensors.push((li, format!("{name} (missing)"))),
-            }
-        }
-    }
-
-    let vmm_probe = inspect_bake_vmm_probe(&store, text_config, weight_prefix, ordinal);
-
-    Some(BakeAccount {
-        bake_dir,
-        manifest_format_version: manifest.format_version,
-        manifest_converter_version: manifest.converter_version,
-        tensor_count: manifest.tensors.len(),
-        weights_bin_bytes,
-        by_layout,
-        missing_specs,
-        bad_expert_tensors,
-        int4_quantized_bytes,
-        raw_bytes,
-        vmm_probe,
-    })
-}
-
-fn inspect_bake_vmm_probe(
-    store: &BakedStore,
-    text_config: &TextConfig,
-    weight_prefix: &str,
-    ordinal: usize,
-) -> Option<VirtualBakeProbe> {
-    let weight_probe = std::env::var_os("SUPERSONIC_VMM_WEIGHT_PROBE").is_some();
-    let moe_probe_layers = std::env::var("SUPERSONIC_VMM_MOE_ISLAND_PROBE_LAYERS")
-        .ok()
-        .and_then(|value| value.parse::<usize>().ok())
-        .or_else(|| std::env::var_os("SUPERSONIC_VMM_MOE_ISLAND_PROBE").map(|_| 1))
-        .unwrap_or(0);
-    if !weight_probe && moe_probe_layers == 0 {
-        return None;
-    }
-    if !gpu_hal::vmm_is_supported(gpu_hal::current_backend(), ordinal) {
-        return None;
-    }
-
-    let mut arena = BakedStore::virtual_weight_arena(ordinal);
-    let mut loaded_tensors = Vec::new();
-    let mut failed_tensors = Vec::new();
-    if weight_probe {
-        let name = resolve_qwen36_store_name(store, "lm_head.weight");
-        match store.load_to_virtual_arena(&mut arena, name.as_ref(), VirtualAllocationRole::Weights)
-        {
-            Ok(_) => loaded_tensors.push(name.into_owned()),
-            Err(err) => failed_tensors.push(format!("{name}: {err}")),
-        }
-    }
-
-    for li in 0..moe_probe_layers.min(text_config.num_hidden_layers as usize) {
-        for kind in ["gate_up_proj", "down_proj"] {
-            let requested = format!("{weight_prefix}.layers.{li}.mlp.experts.{kind}");
-            let name = resolve_qwen36_store_name(store, &requested);
-            match store.load_to_virtual_arena(
-                &mut arena,
-                name.as_ref(),
-                VirtualAllocationRole::MoeExpert,
-            ) {
-                Ok(_) => loaded_tensors.push(name.into_owned()),
-                Err(err) => failed_tensors.push(format!("{name}: {err}")),
-            }
-        }
-    }
-
-    if loaded_tensors.is_empty() && failed_tensors.is_empty() {
-        return None;
-    }
-    let stats = arena.stats();
-    Some(VirtualBakeProbe {
-        allocations: stats.allocations,
-        logical_bytes: stats.logical_bytes,
-        reserved_bytes: stats.reserved_bytes,
-        resident_bytes: stats.resident_bytes,
-        logical_resident_bytes: stats.logical_resident_bytes,
-        mapping_count: stats.mapping_count,
-        loaded_tensors,
-        failed_tensors,
-    })
-}
-
-fn no_bake_only_safetensors(_model_dir: &Path, no_bake: bool) -> bool {
-    // PR 3 only supports safetensors. Future PRs will route to the baked
-    // store when `no_bake` is false and a bake is on disk.
-    let _ = no_bake;
-    true
-}
-
-fn has_safetensors(model_dir: &Path) -> bool {
-    model_dir.join("model.safetensors.index.json").exists()
-        || model_dir.join("model.safetensors").exists()
-}
-
-fn scalar_to_checkpoint_dtype(s: ScalarKind) -> Option<CheckpointDtype> {
-    match s {
-        ScalarKind::Bf16 => Some(CheckpointDtype::Bf16),
-        ScalarKind::F32 => Some(CheckpointDtype::F32),
-        _ => None,
-    }
-}
-
-fn sanity_check_kernel_params(config: &TextConfig, params: &Qwen36MoeKernelParams) -> Result<()> {
-    if config.num_experts as u32 > params.num_experts {
-        return Err(anyhow!(
-            "config has num_experts={} but registry kernel scratch is sized for at most {}; \
-             update Qwen36MoeKernelParams or use a smaller checkpoint",
-            config.num_experts,
-            params.num_experts,
-        ));
-    }
-    if config.num_experts_per_tok as u32 > params.top_k {
-        return Err(anyhow!(
-            "config has num_experts_per_tok={} but registry kernel scratch is sized for top-{}; \
-             update Qwen36MoeKernelParams or use a smaller checkpoint",
-            config.num_experts_per_tok,
-            params.top_k,
-        ));
-    }
-    if config.moe_intermediate_size as u32 > params.moe_intermediate_size {
-        return Err(anyhow!(
-            "config moe_intermediate_size={} exceeds registry bound {}",
-            config.moe_intermediate_size,
-            params.moe_intermediate_size,
-        ));
-    }
-    if config.shared_expert_intermediate_size as u32 > params.shared_expert_intermediate_size {
-        return Err(anyhow!(
-            "config shared_expert_intermediate_size={} exceeds registry bound {}",
-            config.shared_expert_intermediate_size,
-            params.shared_expert_intermediate_size,
-        ));
-    }
-    Ok(())
-}
-
-pub fn print_report(report: &DryRunReport) {
-    let cfg = &report.config.text_config;
-    println!("[qwen3.6-moe] dry-run summary");
-    println!(
-        "  arch:           hidden={} layers={} (full={} linear={}) Q/KV heads={}/{} head_dim={}",
-        cfg.hidden_size,
-        cfg.num_hidden_layers,
-        cfg.num_full_attention_layers(),
-        cfg.num_linear_attention_layers(),
-        cfg.num_attention_heads,
-        cfg.num_key_value_heads,
-        cfg.head_dim,
-    );
-    println!(
-        "  moe:            num_experts={} top_k={} moe_int={} shared_int={} norm_topk={} attn_output_gate={}",
-        cfg.num_experts,
-        cfg.num_experts_per_tok,
-        cfg.moe_intermediate_size,
-        cfg.shared_expert_intermediate_size,
-        cfg.norm_topk_prob,
-        cfg.attn_output_gate,
-    );
-    println!(
-        "  vocab:          {}    rope_theta={}    max_pos={}    tie_embed={}",
-        cfg.vocab_size,
-        cfg.rope_theta(),
-        cfg.max_position_embeddings,
-        cfg.tie_word_embeddings,
-    );
-    let kp = &report.kernel_params;
-    println!(
-        "  kernel params:  prefix={} kv_chunk={} proj_buf_f={} attn_scratch_f={} moe_scratch_f={}",
-        kp.weight_prefix,
-        kp.kv_chunk_size,
-        kp.proj_buf_floats,
-        kp.attn_scratch_floats,
-        kp.moe_scratch_floats,
-    );
-    println!();
-    println!("[checkpoint accounting (analytic, BF16 + F32 mix)]");
-    println!(
-        "  embed:          {:>8.2} GiB",
-        report.checkpoint.embed_bytes as f64 / GIB
-    );
-    println!(
-        "  lm_head:        {:>8.2} GiB",
-        report.checkpoint.lm_head_bytes as f64 / GIB
-    );
-    println!(
-        "  full attn:      {:>8.2} GiB ({} layers × {:.1} MiB)",
-        (report.checkpoint.num_full_layers * report.checkpoint.full_attn_bytes_per_layer) as f64
-            / GIB,
-        report.checkpoint.num_full_layers,
-        report.checkpoint.full_attn_bytes_per_layer as f64 / MIB,
-    );
-    println!(
-        "  linear attn:    {:>8.2} GiB ({} layers × {:.1} MiB)",
-        (report.checkpoint.num_linear_layers * report.checkpoint.linear_attn_bytes_per_layer)
-            as f64
-            / GIB,
-        report.checkpoint.num_linear_layers,
-        report.checkpoint.linear_attn_bytes_per_layer as f64 / MIB,
-    );
-    let n_hidden = cfg.num_hidden_layers as u64;
-    println!(
-        "  routers:        {:>8.2} GiB ({} layers × {:.1} MiB)",
-        (n_hidden * report.checkpoint.router_bytes_per_layer) as f64 / GIB,
-        n_hidden,
-        report.checkpoint.router_bytes_per_layer as f64 / MIB,
-    );
-    println!(
-        "  shared experts: {:>8.2} GiB ({} layers × {:.1} MiB)",
-        (n_hidden * report.checkpoint.shared_expert_bytes_per_layer) as f64 / GIB,
-        n_hidden,
-        report.checkpoint.shared_expert_bytes_per_layer as f64 / MIB,
-    );
-    println!(
-        "  routed experts: {:>8.2} GiB ({} layers × {:.1} GiB / {} experts)",
-        (n_hidden * report.checkpoint.experts_bytes_per_layer) as f64 / GIB,
-        n_hidden,
-        report.checkpoint.experts_bytes_per_layer as f64 / GIB,
-        cfg.num_experts,
-    );
-    println!(
-        "  TOTAL (BF16):   {:>8.2} GiB",
-        report.checkpoint.total_bytes as f64 / GIB
-    );
-    println!(
-        "  TOTAL (INT4):   {:>8.2} GiB  (projected, gs=128)",
-        report.int4_projected_bytes as f64 / GIB
-    );
-    println!();
-    println!("[state accounting]");
-    let max_pos = report.config.text_config.max_position_embeddings;
-    let ctx_label = match report.context_size_source {
-        ContextSizeSource::Explicit => "explicit --context-size".to_string(),
-        ContextSizeSource::EstimatedFromPrompt => {
-            format!("estimated from --prompt (chars + max_new_tokens)")
-        }
-        ContextSizeSource::MaxNewTokensOnly => "max_new_tokens only".to_string(),
-    };
-    println!(
-        "  context tokens:     {:>8}     (source: {ctx_label}; model max_pos={max_pos})",
-        report.context_size_used,
-    );
-    println!(
-        "  KV cache (full attn): {:>8.2} GiB",
-        report.state.full_kv_bytes as f64 / GIB,
-    );
-    println!(
-        "  linear conv state:  {:>8.2} MiB",
-        report.state.linear_conv_state_bytes as f64 / MIB
-    );
-    println!(
-        "  linear recurrent:   {:>8.2} MiB",
-        report.state.linear_recurrent_state_bytes as f64 / MIB
-    );
-    println!(
-        "  moe scratch:        {:>8.2} KiB",
-        report.state.moe_scratch_bytes as f64 / 1024.0
-    );
-    println!(
-        "  activations:        {:>8.2} KiB",
-        report.state.activation_bytes as f64 / 1024.0
-    );
-    println!(
-        "  TOTAL state:        {:>8.2} GiB",
-        report.state.total_bytes as f64 / GIB
-    );
-    println!();
-    if let (Some(on_disk), Some(count)) = (report.on_disk_bytes, report.on_disk_tensor_count) {
-        println!("[on-disk safetensors]");
-        println!("  tensor count:    {count}");
-        println!(
-            "  total bytes:     {:.2} GiB  (analytic predicted {:.2} GiB; drift {:+.2} MiB)",
-            on_disk as f64 / GIB,
-            report.checkpoint.total_bytes as f64 / GIB,
-            (on_disk as f64 - report.checkpoint.total_bytes as f64) / MIB,
-        );
-        if !report.missing_tensors.is_empty() {
-            println!(
-                "  MISSING TENSORS: {} (showing first 10)",
-                report.missing_tensors.len()
-            );
-            for n in report.missing_tensors.iter().take(10) {
-                println!("    - {n}");
-            }
-        }
-        if !report.dtype_mismatches.is_empty() {
-            println!(
-                "  DTYPE MISMATCHES: {} (showing first 10)",
-                report.dtype_mismatches.len()
-            );
-            for n in report.dtype_mismatches.iter().take(10) {
-                println!("    - {n}");
-            }
-        }
-    } else {
-        println!("[on-disk safetensors] not present at model-dir; analytic numbers only");
-    }
-    if let Some(w) = &report.loader_warning {
-        println!("  WARNING: {w}");
-    }
-    println!();
-    if let Some(bake) = &report.bake {
-        println!("[INT4 baked package]");
-        println!("  bake_dir:        {}", bake.bake_dir.display());
-        println!(
-            "  manifest:        format_version={} converter_version={}",
-            bake.manifest_format_version, bake.manifest_converter_version,
-        );
-        println!(
-            "  weights.bin:     {:.2} GiB    ({} tensors indexed)",
-            bake.weights_bin_bytes as f64 / GIB,
-            bake.tensor_count,
-        );
-        println!(
-            "  INT4 / Raw:      {:.2} GiB / {:.2} GiB",
-            bake.int4_quantized_bytes as f64 / GIB,
-            bake.raw_bytes as f64 / GIB,
-        );
-        let parts: Vec<String> = bake
-            .by_layout
-            .iter()
-            .map(|(l, n)| format!("{l}={n}"))
-            .collect();
-        println!("  layouts:         {}", parts.join(", "));
-        // Compare against the analytic INT4 projection — small drift is
-        // expected (the analytic model rounds shapes; the bake is exact).
-        let analytic_gib = report.int4_projected_bytes as f64 / GIB;
-        let bake_gib = bake.weights_bin_bytes as f64 / GIB;
-        println!(
-            "  vs analytic:     bake={:.2} GiB analytic_int4={:.2} GiB drift={:+.2} GiB",
-            bake_gib,
-            analytic_gib,
-            bake_gib - analytic_gib,
-        );
-        if !bake.missing_specs.is_empty() {
-            println!(
-                "  MISSING SPECS:   {} (showing first 10)",
-                bake.missing_specs.len()
-            );
-            for n in bake.missing_specs.iter().take(10) {
-                println!("    - {n}");
-            }
-        }
-        if !bake.bad_expert_tensors.is_empty() {
-            println!(
-                "  BAD EXPERT TENSORS: {} (showing first 10)",
-                bake.bad_expert_tensors.len()
-            );
-            for (li, n) in bake.bad_expert_tensors.iter().take(10) {
-                println!("    - layer {li}: {n}");
-            }
-        }
-        let bake_ok = bake.missing_specs.is_empty() && bake.bad_expert_tensors.is_empty();
-        println!("  ready-for-decode: {}", if bake_ok { "YES" } else { "NO" });
-        if let Some(vmm) = &bake.vmm_probe {
-            println!();
-            println!("  [VMM residency probe]");
-            println!(
-                "    allocations:      {}    mappings={}",
-                vmm.allocations, vmm.mapping_count
-            );
-            println!(
-                "    logical/resident: {:.2} MiB / {:.2} MiB",
-                vmm.logical_bytes as f64 / MIB,
-                vmm.resident_bytes as f64 / MIB,
-            );
-            println!(
-                "    logical resident: {:.2} MiB    reserved={:.2} MiB",
-                vmm.logical_resident_bytes as f64 / MIB,
-                vmm.reserved_bytes as f64 / MIB,
-            );
-            for name in vmm.loaded_tensors.iter().take(8) {
-                println!("    - {name}");
-            }
-            if vmm.loaded_tensors.len() > 8 {
-                println!("    ... {} more", vmm.loaded_tensors.len() - 8);
-            }
-            if !vmm.failed_tensors.is_empty() {
-                println!(
-                    "    failed loads:    {} (showing first 8)",
-                    vmm.failed_tensors.len()
-                );
-                for name in vmm.failed_tensors.iter().take(8) {
-                    println!("    ! {name}");
-                }
-            }
-        }
-        println!();
-    } else {
-        println!(
-            "[INT4 baked package] not present at .supersonic/v{}-int4-gptq/ \
-             (run `oracle/bake_int4.py` to produce one)",
-            model_store::manifest::FORMAT_VERSION
-        );
-        println!();
-    }
-    println!("[VRAM budget]");
-    println!(
-        "  registry estimate (INT4 weights + KV @ ctx, ×overhead): {:.2} GiB",
-        report.registry_budget_bytes as f64 / GIB
-    );
-    println!(
-        "  detected GPU VRAM:                                       {:.2} GiB",
-        report.gpu_total_vram_bytes as f64 / GIB
-    );
-    let fits = report.registry_budget_bytes <= report.gpu_total_vram_bytes;
-    println!(
-        "  fit:                                                     {}",
-        if fits { "YES" } else { "NO" }
-    );
-    if !fits {
-        println!(
-            "  (BF16 weights alone need ~{:.1} GiB; an INT4/q4km bake is required.)",
-            report.checkpoint.total_bytes as f64 / GIB
-        );
-    }
-    if report.context_size_source == ContextSizeSource::MaxNewTokensOnly {
-        println!();
-        println!(
-            "  WARNING: context_size defaulted to max_new_tokens={}. KV cache for a real",
-            report.context_size_used,
-        );
-        println!("  decode session will be larger; pass --context-size <N> (or --prompt) for an",);
-        println!(
-            "  honest fit number. Worst case at the model's max ({} tokens) needs",
-            max_pos,
-        );
-        // Quick worst-case KV estimate at model max — gives the user one
-        // concrete data point before they have to re-run with a flag.
-        let worst_kv = report.config.text_config.kv_bytes_per_token(2) * max_pos as u64;
-        println!(
-            "  ~{:.2} GiB of KV cache alone (BF16).",
-            worst_kv as f64 / GIB,
-        );
-    }
-}
 
 pub fn run(cli: &crate::Cli, entry: &RegistryEntry, total_vram: u64) -> Result<()> {
     ensure_qwen36_bake(cli, entry)?;
